Moving ethSign benchmarking to Local Runner
