+++ rskj-core/src/jmh/java/co/rsk/jmh/web3/BenchmarkLocalWalletWeb3.java
@@ -18,6 +18,7 @@
 
 package co.rsk.jmh.web3;
 
+import co.rsk.jmh.web3.plan.LocalPlan;
 import co.rsk.jmh.web3.plan.TransactionPlan;
 import org.openjdk.jmh.annotations.*;
 import org.web3j.protocol.core.methods.request.Transaction;
@@ -53,6 +54,13 @@ public class BenchmarkLocalWalletWeb3 {
     public void ethSendTransaction_ContractCall(TransactionPlan plan) throws BenchmarkWeb3Exception {
         Transaction tx = plan.getTransactionsContractCreation().next();
         plan.getWeb3Connector().ethSendTransaction(tx);
+    }
+
+    @Benchmark
+    public void ethSign(LocalPlan plan) throws BenchmarkWeb3Exception {
+        String address = plan.getEthSignAddress();
+        String message = plan.getEthSignMessage();
+        plan.getWeb3Connector().ethSign(address, message);
     }
 
 }
