+++ rskj-core/src/jmh/java/co/rsk/jmh/web3/EthMethodsConfig.java
@@ -26,24 +26,21 @@ public class EthMethodsConfig {
     public static final String ETH_GET_STORAGE_AT_POSITION = "ethGetStorageAt.position";
     public static final String ETH_GET_CODE_ADDRESS = "ethGetCode.address";
     public static final String ETH_GET_CODE_BLOCK_NUMBER = "ethGetCode.blockNumber";
-    public static final String ETH_SIGN_ADDRESS = "ethSign.address";
-    public static final String ETH_SIGN_MESSAGE = "ethSign.message";
+
 
 
     private final String ethGetStorageAtAddress;
     private final Long ethGetStorageAtPosition;
     private final String ethGetCodeAddress;
     private final Long ethGetCodeBlockNumber;
-    private final String ethSignAddress;
-    private final String ethSignMessage;
+
 
     public EthMethodsConfig(Config config) {
         ethGetStorageAtAddress = config.getString(ETH_GET_STORAGE_AT_ADDRESS);
         ethGetStorageAtPosition = config.getLong(ETH_GET_STORAGE_AT_POSITION);
         ethGetCodeAddress = config.getString(ETH_GET_CODE_ADDRESS);
         ethGetCodeBlockNumber = config.getLong(ETH_GET_CODE_BLOCK_NUMBER);
-        ethSignAddress = config.getString(ETH_SIGN_ADDRESS);
-        ethSignMessage = config.getString(ETH_SIGN_MESSAGE);
+
     }
 
     public String getEthGetStorageAtAddress() {
@@ -66,11 +63,4 @@ public class EthMethodsConfig {
         return DefaultBlockParameter.valueOf("latest");
     }
 
-    public String getEthSignAddress() {
-        return ethSignAddress;
-    }
-
-    public String getEthSignMessage() {
-        return ethSignMessage;
-    }
 }
