+++ rskj-core/src/jmh/java/co/rsk/jmh/web3/plan/BasePlan.java
@@ -95,14 +95,6 @@ public class BasePlan {
         return config;
     }
 
-    public BenchmarkWeb3.Suites getSuite() {
-        return suite;
-    }
-
-    public String getHost() {
-        return host;
-    }
-
     public EthMethodsConfig getEthMethodsConfig() {
         return ethMethodsConfig;
     }
