+++ rskj-core/src/jmh/resources/conf/testnet-3_860_000.conf
@@ -53,8 +53,6 @@ ethGetStorageAt.address = 0x19f64674D8a5b4e652319F5e239EFd3bc969a1FE
 ethGetStorageAt.position = 1
 ethGetCode.address = 0x462d7082F3671a3be160638Be3F8c23cA354f48a
 ethGetCode.blockNumber = 3700000
-ethSign.address = 0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826
-ethSign.message = 0xdeadbeaf
 
 ############################
 # trace
